Add util::map_ptr to incFs
util::map_ptr is a smart pointer that is able to verify whether the
contents of the pointer are fully present on the file system before
the pointer is used. The pointer points to memory-mapped data
owned by util::IncFsFileMap.

This utility class allows libraries that may memory-mapped incremental
files to continue using pointer arithmetic with minimal refactoring of
code logic and allows them to handle missing sections of the files
without the possibility of SIGBUS signals being raised.

util::IncFsFileMap has a DEBUG flag that is used to verify that the
contents of pointers are always verified before the pointer is
de-referenced.

Raw pointers can be implicitly casted to util::map_ptr and have
negligible overhead.

The pointer assumes that:
1) A memory-mapped page is fully present on IncFs if an invocation of
   pread on the data of the page does not fail.
2) Any data region that exists entirely within fully present pages is
   also fully present.

Bug: 160635104
Bug: 168064232
Test: atest libincf-test
Merged-In: I2e96538afc1344f9d83a65a37f7164273cfbad6e
Change-Id: I2e96538afc1344f9d83a65a37f7164273cfbad6e