Fix map_ptr "Bad file descriptor"
It is common to close a file descriptor after calling mmap since the
map stays open until munmap is invoked. The current implementation
of IncfsFileMap checks if block verification is enabled each time the
pointer to the mmaped data is retrieved. Instead, only check if
verification is enabled when IncfsFileMap::Create is invoked while the
fd is guaranteed to be open.

Bug: 173254631
Test: boot and verify that no more "Bad file descriptor" logs occur
Test: libincfs-test
Change-Id: Icddb0c9af717927ef8f186b64ac0287ce0ba4acd